requirements.txt
pandas>=1.5.0
requests>=2.28.0
Pillow>=10.0.0
firebase-admin>=6.2.0 